feat: prompt for start block on rpc url networks in contract import (#4)
* feat: prompt for start_block on rpc contract import indexers

* tests

* feat: use custom type to allow recoverability for invalid startblock

---------

+++ codegenerator/cli/src/config_parsing/chain_helpers.rs
@@ -343,7 +343,7 @@ impl Network {
     }
 
     /// Returns the end block for this network if it is finite
-    pub fn get_finite_end_block(&self) -> Option<i32> {
+    pub fn get_finite_end_block(&self) -> Option<u64> {
         match self {
             Self::Goerli => Some(10_387_962),
             Self::Mumbai => Some(47_002_303),
